Refactored: renanmed NavbarComponents to follow camelCase standard
